// src/protocol/message-types.ts
export enum MessageType {
  LOGIN = "LOGIN",
  LOGIN_RESPONSE = "LOGIN_RESPONSE",

  LOGOUT = "LOGOUT",
  LOGOUT_RESPONSE = "LOGOUT_RESPONSE",

  GET_GROUPS = "GET_GROUPS",
  GET_GROUPS_RESPONSE = "GET_GROUPS_RESPONSE",

  CREATE_GROUP = "CREATE_GROUP",
  CREATE_GROUP_RESPONSE = "CREATE_GROUP_RESPONSE",

  JOIN_GROUP = "JOIN_GROUP",
  JOIN_GROUP_RESPONSE = "JOIN_GROUP_RESPONSE",

  LEAVE_GROUP = "LEAVE_GROUP",
  LEAVE_GROUP_RESPONSE = "LEAVE_GROUP_RESPONSE",

  ADD_MEMBER = "ADD_MEMBER",
  ADD_MEMBER_RESPONSE = "ADD_MEMBER_RESPONSE",

  REMOVE_MEMBER = "REMOVE_MEMBER",
  REMOVE_MEMBER_RESPONSE = "REMOVE_MEMBER_RESPONSE",

  GET_MESSAGES = "GET_MESSAGES",
  GET_MESSAGES_RESPONSE = "GET_MESSAGES_RESPONSE",

  SEND_MESSAGE = "SEND_MESSAGE",
  SEND_MESSAGE_RESPONSE = "SEND_MESSAGE_RESPONSE",

  NEW_MESSAGE = "NEW_MESSAGE",

  PING = "PING",
  PONG = "PONG",

  ERROR = "ERROR"
}